# API.yaml
openapi: 3.0.3
info:
  title: Printer Queue Generation API
  description: |
    API para geração de filas de impressão com empacotamento automático de imagens.
    
    ## Funcionalidades Principais
    
    - **Gestão de Tenants**: Multi-tenancy com isolamento de dados
    - **Gestão de Máquinas**: Configuração de máquinas de impressão por tenant
    - **Gestão de Assets**: Catálogo de imagens indexadas por SKU
    - **Gestão de Perfis de Dimensionamento**: Perfis para dimensionamento automático de imagens
    - **Gestão de Storage**: Configuração de storage (S3, Local, Dropbox)
    - **Processamento de Jobs**: Upload de picklists PDF, resolução de SKUs, empacotamento e geração de PDFs
    - **Image Packing**: Upload direto de imagens para empacotamento sem validação de SKU
    
    ## Autenticação
    
    A API usa multi-tenancy baseado em header `X-Tenant-ID`. Todos os endpoints (exceto `/healthz`) requerem este header.
    
    ## Status de Jobs
    
    - `queued`: Job criado, aguardando processamento
    - `processing`: Job sendo processado pelo worker
    - `needs_input`: Job precisa de resolução manual de SKUs
    - `completed`: Job processado com sucesso
    - `failed`: Job falhou durante processamento
    
    ## Modos de Processamento
    
    - `sequence`: Processa itens na ordem exata do picklist PDF
    - `optimize`: Otimiza layout para minimizar número de bases e desperdício
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
...
paths:
...
  /v1/image-packing/upload:
    post:
      tags:
        - image-packing
      summary: Upload images for automatic packing
      description: |
        Upload images or ZIP file for automatic packing/layout generation.
        
        **Key Features:**
        - No SKU validation required
        - No sizing profiles needed
        - Accepts multiple image files or a single ZIP archive
        - Automatically extracts images from ZIP
        - Calculates dimensions from image metadata (DPI, pixels)
        - Generates best possible layout to minimize bases and waste
        - Respects machine dimensions and cut margins
        
        **Supported Formats:**
        - PNG, JPEG, GIF, WebP, BMP, TIFF
        - ZIP archives containing images
        
        **Processing:**
        - Images are processed asynchronously
        - Job status can be checked via `/status/{job_id}`
        - Result can be downloaded as multi-page PDF via `/download/{job_id}`
      operationId: uploadImages
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - files
                - machine_width_mm
                - machine_length_mm
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Image files or ZIP archive
                mode:
                  type: string
                  enum: [sequence, optimize]
                  default: optimize
                  description: Packing mode
                machine_width_mm:
                  type: number
                  format: float
                  description: Machine width in millimeters
                machine_length_mm:
                  type: number
                  format: float
                  description: Machine length in millimeters
      responses:
        '201':
          description: Images uploaded, processing started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImagePackingUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/image-packing/status/{job_id}:
    get:
      tags:
        - image-packing
      summary: Get image packing job status
      description: Get current status and progress of an image packing job
      operationId: getImagePackingStatus
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImagePackingStatusResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/image-packing/result/{job_id}:
    get:
      tags:
        - image-packing
      summary: Get image packing result
      description: Get packing result (bases, placements, utilization) for completed jobs
      operationId: getImagePackingResult
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Packing result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImagePackingResultResponse'
        '400':
          description: Job not completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/image-packing/download/{job_id}:
    get:
      tags:
        - image-packing
      summary: Download packing result as PDF
      description: Download generated layout as a multi-page PDF (one page per base)
      operationId: downloadImagePackingResult
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Multi-page PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          description: Job not completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
    TenantIdHeader:
      name: X-Tenant-ID
      in: header
      required: true
      description: Tenant ID for multi-tenancy
      schema:
        type: integer
...
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message
...
    ImagePackingUploadResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued]
        message:
          type: string
    ImagePackingStatusResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, processing, completed, failed]
        progress:
          type: integer
          nullable: true
          minimum: 0
          maximum: 100
        message:
          type: string
          nullable: true
        result:
          $ref: '#/components/schemas/ImagePackingResultResponse'
          nullable: true
        error:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ImagePackingResultResponse:
      type: object
      properties:
        total_bases:
          type: integer
        total_length_mm:
          type: number
          format: float
        avg_utilization:
          type: number
          format: float
        mode:
          type: string
          enum: [sequence, optimize]
        bases:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              width_mm:
                type: number
                format: float
              length_mm:
                type: number
                format: float
              utilization:
                type: number
                format: float
              items_count:
                type: integer
              placements:
                  type: array
                  items:
                    type: object
                    properties:
                      item_id:
                        type: integer
                      sku:
                        type: string
                      x_mm:
                        type: number
                        format: float
                      y_mm:
                        type: number
                        format: float
                      width_mm:
                        type: number
                        format: float
                      height_mm:
                        type: number
                        format: float
                      rotated:
                        type: boolean
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'